requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
entrypoints==0.4
faiss-cpu==1.7.4
Faker==37.4.0
favicon==0.7.0
filelock==3.18.0
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.43.0
numpy==2.3.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
prometheus_client==0.22.1
protobuf==5.29.5
pyarrow==18.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.15
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-snowpark-python==1.32.0
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.45.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.1
streamlit-faker==0.0.3
streamlit-image-coordinates==0.3.1
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==9.1.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
validators==0.35.0
zstandard==0.23.0
